// src/types/generated/PromptParameter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PromptParameterType } from "./PromptParameterType";

export interface PromptParameter { id: number | null, versionId: number, key: string, value: string, parameterType: PromptParameterType, description: string | null, }